refactor: trading과 관련된 코드들의 패키지 위치가 불분명하여 수정
