+++ domain/src/main/java/com/fintory/domain/portfolio/service/ExchangeRateService.java
@@ -1,4 +1,4 @@
-package com.fintory.domain.stock.service;
+package com.fintory.domain.portfolio.service;
 
 import java.math.BigDecimal;
